+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- entries api: use POST based searches if the browser is IE and the url length is greater than 2083.
+- entries api: new distanceWithin operator for location based searches.
+
+## [1.1.0-beta.6] 2019-03-26
+### Added
+- entries api: changed the request type from POST to GET for searches.
+
 ## [1.1.0-beta.5] 2019-03-26
 ### Changed
 - nodes api: enhanced api error processing.